VIITE-84 Refactor, extract project service out of road address service, fix tests
