+++ src/main/scala/ScalatraBootstrap.scala
@@ -27,7 +27,8 @@ ScalatraBootstrap extends LifeCycle {
     context.mount(new IntegrationApi(Digiroad2Context.massTransitStopService), "/api/integration/*")
     context.mount(new ViiteIntegrationApi(Digiroad2Context.roadAddressService), "/api/viite/integration/*")
     context.mount(new ChangeApi(), "/api/changes/*")
-    context.mount(new ViiteApi(Digiroad2Context.roadLinkService, Digiroad2Context.vvhClient, Digiroad2Context.roadAddressService), "/api/viite/*")
+    context.mount(new ViiteApi(Digiroad2Context.roadLinkService, Digiroad2Context.vvhClient,
+      Digiroad2Context.roadAddressService, Digiroad2Context.projectService), "/api/viite/*")
     if (!Digiroad2Context.getProperty("digiroad2.tierekisteri.enabled").toBoolean) {
       context.mount(new TierekisteriTestApi, "/api/tierekisteri/*")
     }
